% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SecAct_wrappers.R
\name{SecAct.inference.gsl.new}
\alias{SecAct.inference.gsl.new}
\title{Secreted protein activity inference (Optimized Version)}
\usage{
SecAct.inference.gsl.new(
  Y,
  SigMat = "SecAct",
  lambda = 5e+05,
  nrand = 1000,
  ncores = NULL
)
}
\arguments{
\item{Y}{Gene expression matrix.}

\item{SigMat}{Secreted protein signature matrix.}

\item{lambda}{Penalty factor.}

\item{nrand}{Number of randomizations.}

\item{ncores}{Number of cores (default: all available).}
}
\description{
Fast multi-threaded implementation using OpenMP and .Call interface for large data.
}
