% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SecAct_wrappers.R
\name{SecAct.inference.gsl.old}
\alias{SecAct.inference.gsl.old}
\title{Secreted protein activity inference (Legacy Version)}
\usage{
SecAct.inference.gsl.old(Y, SigMat = "SecAct", lambda = 5e+05, nrand = 1000)
}
\arguments{
\item{Y}{Gene expression matrix.}

\item{SigMat}{Secreted protein signature matrix.}

\item{lambda}{Penalty factor.}

\item{nrand}{Number of randomizations.}
}
\description{
Old single-threaded implementation.
}
